[Enhancements] [CCNET] [CCNET Version Upgrade] Modified/migrated some functionality in the CI Factory CCNET source from the mainline CCNET source (CCNET v1.5). 1. HtmlReportPlugin: Added the HtmlReportPlugin (plus several other classes from CCNET v1.5 that it depends upon). 2. XsltParameter: Added the Actions/XsltParameter class.
NOTE: The FingerprintFactory.BuildFromRequest method was modified during the migration. It is effectively still NOT implemented in the CI Factory source, because the CCNET v1.5 implementation relies upon multiple API changes to some underlying base classes. However, that method is not currently called by any CI Factory code, so leaving it with a "not implemented" implementation should have no negative effects for now.