Enable nAnt scripts to pass parameters to groovy task definitions
Currently, the task scans the original groovy source code to determine what the parameter names are. We'll
need to replace this functionality with something more robust (either scanning the abstract syntax tree or
getting a third party external library like paranamer to work correctly when applied against java bytecode
in the CLR).